backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
alembic>=1.13.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart>=0.0.6
argon2-cffi==23.1.0
pydantic[email]>=2.5.0
python-dotenv>=1.0.0

# ML inference dependencies
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.4.0
xgboost>=2.0.0
shap>=0.44.0
joblib>=1.3.0

# Embryo grading (Phase 3)
Pillow>=10.0.0
torch>=2.1.0
torchvision>=0.16.0
